LayoutTests/imported/w3c/web-platform-tests/fetch/metadata/script.sub-expected.txt

Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/echo-as-script.py
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/echo-as-script.py

PASS Non-secure same-origin script => No headers
FAIL Non-secure same-origin script => No headers: sec-fetch-dest assert_equals: expected "" but got "script"
FAIL Non-secure same-origin script => No headers: sec-fetch-mode assert_equals: expected "" but got "no-cors"
FAIL Non-secure same-origin script => No headers: sec-fetch-site assert_equals: expected "" but got "same-origin"
PASS Non-secure same-origin script => No headers: sec-fetch-user
PASS Non-secure same-site script => No headers
FAIL Non-secure same-site script => No headers: sec-fetch-dest null is not an object (evaluating 'value.dest')
FAIL Non-secure same-site script => No headers: sec-fetch-mode null is not an object (evaluating 'value.mode')
FAIL Non-secure same-site script => No headers: sec-fetch-site null is not an object (evaluating 'value.site')
FAIL Non-secure same-site script => No headers: sec-fetch-user null is not an object (evaluating 'value.user')
PASS Non-secure cross-site script => No headers
FAIL Non-secure cross-site script => No headers: sec-fetch-dest null is not an object (evaluating 'value.dest')
FAIL Non-secure cross-site script => No headers: sec-fetch-mode null is not an object (evaluating 'value.mode')
FAIL Non-secure cross-site script => No headers: sec-fetch-site null is not an object (evaluating 'value.site')
FAIL Non-secure cross-site script => No headers: sec-fetch-user null is not an object (evaluating 'value.user')
